Deleted MainActivity and make Information Activity
