+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,13 @@
         android:theme="@style/Theme.Front_End"
         tools:targetApi="31">
         <activity
+            android:name=".Information"
+            android:exported="false">
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
+        <activity
             android:name=".SightsActivity"
             android:exported="false">
             <meta-data
